requirements.txt
asgiref @ file:///home/lenovo/.cache/pypoetry/artifacts/12/c6/58/b461520efbe0ac0a67261f2d665156c3efabcc6392f021f3db62abbd51/asgiref-3.5.0-py3-none-any.whl
backports.zoneinfo==0.2.1
black==21.12b0
certifi==2021.10.8
charset-normalizer @ file:///home/lenovo/.cache/pypoetry/artifacts/d5/30/7e/2900610df8360e6f3f6493f0fdf18de011be2a4c768ae0ef536861dddb/charset_normalizer-2.0.11-py3-none-any.whl
click==8.0.3
cloudinary==1.28.1
dataclasses==0.6
Django @ file:///home/lenovo/.cache/pypoetry/artifacts/24/48/77/f3e19922053876f74f44b449e8e24b4085e8454c59fb8e7024b2808c0a/Django-4.0.3-py3-none-any.whl
django-cloudinary-storage @ file:///home/lenovo/.cache/pypoetry/artifacts/1e/d9/90/2a393f4d62953de07452314573d82c49e2ade987393400bc6bc9b07551/django_cloudinary_storage-0.3.0-py3-none-any.whl
django-cors-headers @ file:///home/lenovo/.cache/pypoetry/artifacts/25/5b/9f/bc22a92ae6e0ed246fdfd50bf9d0ea31f05f82bce8414ba8c617cff2cf/django_cors_headers-3.11.0-py3-none-any.whl
django-environ @ file:///home/lenovo/.cache/pypoetry/artifacts/48/81/25/27d1f75f11e5ecbf4bc3899a97844d70d8db6ae138fdd42a9a635ba426/django_environ-0.8.1-py2.py3-none-any.whl
django-filter @ file:///home/lenovo/.cache/pypoetry/artifacts/5a/68/8a/77b40dc782662c89a85de68752ca22242c38218dc94ea9115eaa597532/django_filter-21.1-py3-none-any.whl
django-jsonfield @ file:///home/lenovo/.cache/pypoetry/artifacts/1c/0e/a0/96bbd17cc83545848067a08b1f79eb1e9af38ecf24b34def4ac0e36a0f/django_jsonfield-1.4.1-py2.py3-none-any.whl
django-rest-auth==0.9.5
django-stripe-payments==2.0.0
django-uuidfield==0.5.0
django-versatileimagefield @ file:///home/lenovo/.cache/pypoetry/artifacts/f6/30/6f/6d6f6a7baf8843a44a8fb1fa7a9cf612a3bded4d23d31d2f0b40c95ab1/django_versatileimagefield-2.2-py2.py3-none-any.whl
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djongo==1.3.6
dnspython==2.1.0
drf-flex-fields==0.9.7
gunicorn==20.1.0
idna==3.3
mypy-extensions==0.4.3
mysqlclient==2.1.0
pathspec==0.9.0
Pillow==9.0.0
pkg_resources==0.0.0
platformdirs==2.4.1
PyJWT==2.3.0
pymongo==3.12.1
python-magic @ file:///home/lenovo/.cache/pypoetry/artifacts/45/f2/93/ce97a3ad361da42a711d1e85f9bd35de32a7302d4c0cfde99ab6ca9eca/python_magic-0.4.15-py2.py3-none-any.whl
pytz==2021.3
razorpay @ file:///home/lenovo/.cache/pypoetry/artifacts/be/af/73/b258621fef3e248c26a6943276a8b1023492468b680738e42dbfda92a7/razorpay-1.3.0-py3-none-any.whl
requests @ file:///home/lenovo/.cache/pypoetry/artifacts/dc/17/bf/8d9779110a13421f6181028974b24800ec70f28bf9b311411e1868b8c7/requests-2.27.1-py2.py3-none-any.whl
six==1.16.0
sqlparse==0.4.2
stripe @ file:///home/lenovo/.cache/pypoetry/artifacts/b4/43/ae/dfe35817957b4a732e72418a1438d843e42d63170276c074aa0aa8a55d/stripe-2.65.0-py2.py3-none-any.whl
tomli==1.2.3
typing_extensions==4.0.1
urllib3 @ file:///home/lenovo/.cache/pypoetry/artifacts/66/2a/bb/22b56fafb85086d3ee6c6807644cffcd5fa6fc2bc537780143368ab6a9/urllib3-1.26.8-py2.py3-none-any.whl
uuid==1.30
whitenoise @ file:///home/lenovo/.cache/pypoetry/artifacts/b6/4d/0c/bd379f9fa67f591dd8cb8f58707fdaa005992a0329f18c5c9f19a1a0a2/whitenoise-6.0.0-py3-none-any.whl
